Fix bind attributes on relink (ng-if) element #14
This commit fixes the issue that the bind-polymer directive stopped working after the initial link.
This was caused because in the second link, the attribute didn't have the value with the curly braces notation anymore.
To fix this, we retrieve the bind expressions in the compile phase rather than the linking phase.

In this commit I also take the refactorings of @heavysixer (pull request #12) into account.
The watcher is indeed unneeded and it's logic has moved to the MutationObserver.

Also I got rid of the own implementation to traverse expressions by using the Angular $parser.
This returns both a getter and setter used in the MutationObserver.
This probably fixes issue #10 as well.